+++ substra/sdk/client.py
@@ -763,7 +763,7 @@ class Client:
             worker (List[str]): list tasks which ran on listed workers. Remote mode only.\n
             rank (List[int]): list tasks which are at given ranks.\n
             status (List[str]): list tasks with given status.
-                The possible values are the values of `substra.models.Status`
+                The possible values are the values of `substra.models.ComputeTaskStatus`
             metadata (dict)
                 {
                     "key": str # the key of the metadata to filter on
@@ -1062,9 +1062,46 @@ class Client:
                 Not raised when `timeout == None`
         """
         asset_getter = self.get_task
-        status_canceled = models.Status.canceled.value
-        status_failed = models.Status.failed.value
-        statuses_stopped = (models.Status.done.value, models.Status.canceled.value)
+        status_canceled = models.ComputeTaskStatus.canceled.value
+        status_failed = models.ComputeTaskStatus.failed.value
+        statuses_stopped = (models.ComputeTaskStatus.done.value, models.ComputeTaskStatus.canceled.value)
+        return self._wait(
+            key=key,
+            asset_getter=asset_getter,
+            polling_period=polling_period,
+            raise_on_failure=raise_on_failure,
+            status_canceled=status_canceled,
+            status_failed=status_failed,
+            statuses_stopped=statuses_stopped,
+            timeout=timeout,
+        )
+
+    @logit
+    def wait_function(
+        self, key: str, *, timeout: Optional[float] = None, polling_period: float = 1.0, raise_on_failure: bool = True
+    ) -> models.Task:
+        """Wait for the build of the given function to finish.
+
+        It is considered finished when the status is ready, failed or cancelled.
+
+        Args:
+            key (str): the key of the task to wait for.
+            timeout (float, optional): maximum time to wait, in seconds. If set to None, will hang until completion.
+            polling_period (float): time to wait between two checks, in seconds. Defaults to 2.0.
+            raise_on_failure (bool): whether to raise an exception if the execution fails. Defaults to True.
+
+        Returns:
+            models.Task: the task after completion
+
+         Raises:
+            exceptions.FutureFailureError: The task failed or have been cancelled.
+            exceptions.FutureTimeoutError: The task took more than the duration set in the timeout to complete.
+                Not raised when `timeout == None`
+        """
+        asset_getter = self.get_function
+        status_canceled = models.FunctionStatus.canceled.value
+        status_failed = models.FunctionStatus.failed.value
+        statuses_stopped = (models.FunctionStatus.ready.value, models.FunctionStatus.canceled.value)
         return self._wait(
             key=key,
             asset_getter=asset_getter,
@@ -1095,7 +1132,7 @@ class Client:
             if asset.status in statuses_stopped:
                 break
 
-            if asset.status == models.Status.failed.value and asset.error_type is not None:
+            if asset.status == models.ComputeTaskStatus.failed.value and asset.error_type is not None:
                 # when dealing with a failed task, wait for the error_type field of the task to be set
                 # i.e. wait for the registration of the failure report
                 break